+++ src/App.tsx
@@ -17,8 +17,8 @@ import PageNotFound from "./pages/404/404";
 import { getRedirectPathFromParam, tillatteTjenester } from "./utils/redirects";
 import SkattkortHistorikk from "./pages/skattetrekksmelding/SkattHistorikk";
 import SkattekortDetaljer from "./pages/skattetrekksmelding/SkattDetaljer";
-import InstHistorikk from "./pages/institusjonsopphold/InstHistorikk";
-import InstDetaljer from "./pages/institusjonsopphold/InstDetaljer";
+import InstHistorikk from "./pages/institusjonsopphold/historikk/InstHistorikk";
+import InstDetaljer from "./pages/institusjonsopphold/detaljer/InstDetaljer";
 import DsopHistorikkPage from "./pages/digital-samhandling-offentlig-privat/historikk/DsopHistorikkPage";
 import DsopDetaljerPage from "./pages/digital-samhandling-offentlig-privat/detaljer/DsopDetaljerPage";
 import Modal from "react-modal";
